refactor: move folder for common using
