+++ src/main/java/co/orange/ddanzi/controller/ItemController.java
@@ -48,4 +48,9 @@ public class ItemController {
         return itemService.getItem(id);
     }
 
+    @GetMapping("order/{id}")
+    ApiResponse<?> getAddressAndOption(@PathVariable("id") String id) {
+        return itemService.getAddressAndOption(id);
+    }
+
 }
