+++ src/main/java/co/orange/ddanzi/dto/common/AddressSeparateInfo.java
@@ -1,6 +1,5 @@
-package co.orange.ddanzi.dto;
+package co.orange.ddanzi.dto.common;
 
-import lombok.Builder;
 import lombok.Getter;
 import lombok.experimental.SuperBuilder;
