+++ src/main/java/co/orange/ddanzi/dto/item/SaveItemRequestDto.java
@@ -4,11 +4,9 @@ import co.orange.ddanzi.domain.product.Item;
 import co.orange.ddanzi.domain.product.Product;
 import co.orange.ddanzi.domain.product.enums.ItemStatus;
 import co.orange.ddanzi.domain.user.User;
-import co.orange.ddanzi.dto.TermInfo;
 import lombok.Getter;
 
 import java.time.LocalDate;
-import java.util.List;
 
 @Getter
 public class SaveItemRequestDto {
