+++ src/main/java/co/orange/ddanzi/dto/mypage/MyPageInterestResponseDto.java
@@ -1,6 +1,6 @@
 package co.orange.ddanzi.dto.mypage;
 
-import co.orange.ddanzi.dto.ProductInfo;
+import co.orange.ddanzi.dto.common.ProductInfo;
 import lombok.Builder;
 import lombok.Getter;
