+++ src/main/java/co/orange/ddanzi/dto/order/CheckProductResponseDto.java
@@ -1,6 +1,6 @@
 package co.orange.ddanzi.dto.order;
 
-import co.orange.ddanzi.dto.AddressInfo;
+import co.orange.ddanzi.dto.common.AddressInfo;
 import lombok.Builder;
 import lombok.Getter;
