+++ src/main/java/co/orange/ddanzi/service/HomeService.java
@@ -3,7 +3,9 @@ package co.orange.ddanzi.service;
 import co.orange.ddanzi.domain.Banner;
 import co.orange.ddanzi.domain.product.*;
 import co.orange.ddanzi.domain.user.User;
-import co.orange.ddanzi.dto.ProductInfo;
+import co.orange.ddanzi.dto.common.ProductInfo;
+import co.orange.ddanzi.dto.common.OptionDetailInfo;
+import co.orange.ddanzi.dto.common.OptionInfo;
 import co.orange.ddanzi.dto.home.*;
 import co.orange.ddanzi.common.error.Error;
 import co.orange.ddanzi.common.response.ApiResponse;
